be/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
google-generativeai
google-api-core

